# unified-exchange-platform/docs/openapi.yaml
openapi: 3.0.0
info:
  title: KK99 Exchange API
  version: 1.0.0
  description: Production-ready Unified Exchange Platform API
  contact:
    name: KK99 Team
    email: support@kk99.io
servers:
  - url: https://api.exchange.kk99.io/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
paths:
...
  /orders:
    post:
      summary: Create new order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    get:
      summary: Get user orders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: number
        price:
          type: number
        status:
          type: string
        created_at:
          type: string
...
    CreateOrderRequest:
      type: object
      required:
        - symbol
        - side
        - orderType
        - quantity
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        orderType:
          type: string
          enum: [LIMIT, MARKET]
        quantity:
          type: number
        price:
          type: number
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT